+++ src/main/kotlin/com/example/kbocombo/crawler/infrastructure/KboPlayerListPageCrawler.kt
@@ -2,7 +2,6 @@ package com.example.kbocombo.crawler.infrastructure
 
 import com.example.kbocombo.crawler.client.KboHttpClient
 import com.example.kbocombo.crawler.dto.NewPlayerData
-import com.example.kbocombo.crawler.service.PlayerCrawler
 import com.example.kbocombo.domain.player.vo.Team
 import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Component
