+++ src/main/kotlin/com/example/kbocombo/crawler/infrastructure/PlayerCrawler.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.crawler.service
+package com.example.kbocombo.crawler.infrastructure
 
 import com.example.kbocombo.crawler.dto.NewPlayerData
